Refactor MainAppScreen structure
