+++ gradle/libs.versions.toml
@@ -2,10 +2,10 @@
 activityCompose = "1.10.1"
 agp = "8.9.1"
 androidx-hilt-navigation = "1.2.0"
-composeBom = "2024.09.00"
-compose-navigation = "2.8.9"
-coil = "2.5.0"
-coreKtx = "1.15.0"
+composeBom = "2025.06.00"
+compose-navigation = "2.9.0"
+coil = "2.6.0"
+coreKtx = "1.16.0"
 espressoCore = "3.6.1"
 okhttp = "4.12.0"
 ksp = "2.1.0-1.0.29"
@@ -23,6 +23,8 @@ room = "2.7.1"
 [libraries]
 androidx-activity-compose = { group = "androidx.activity", name = "activity-compose", version.ref = "activityCompose" }
 androidx-compose-bom = { group = "androidx.compose", name = "compose-bom", version.ref = "composeBom" }
+androidx-compose-foundation = { group = "androidx.compose.foundation", name = "foundation" }
+androidx-compose-foundation-layout = { group = "androidx.compose.foundation", name = "foundation-layout" }
 androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "coreKtx" }
 androidx-espresso-core = { group = "androidx.test.espresso", name = "espresso-core", version.ref = "espressoCore" }
 androidx-junit = { group = "androidx.test.ext", name = "junit", version.ref = "junitVersion" }
